+++ test/OpenTelemetry.Exporter.Prometheus.Tests/PrometheusCollectionManagerTests.cs
@@ -29,14 +29,22 @@ namespace OpenTelemetry.Exporter.Prometheus.Tests
 {
     public sealed class PrometheusCollectionManagerTests
     {
+#if NETFRAMEWORK
+        [Fact(Skip = "Might be flaky. Might be a bug. See: https://github.com/open-telemetry/opentelemetry-dotnet/pull/3618#issuecomment-1232200946")]
+#else
         [Fact]
+#endif
         public async Task EnterExitCollectTest()
         {
             using var meter = new Meter(Utils.GetCurrentMethodName());
 
             using (var provider = Sdk.CreateMeterProviderBuilder()
                 .AddMeter(meter.Name)
+#if NETFRAMEWORK
+                .AddPrometheusHttpListener()
+#else
                 .AddPrometheusExporter()
+#endif
                 .Build())
             {
                 if (!provider.TryFindExporter(out PrometheusExporter exporter))